multikeyqsort

+++ Bowtie2Java/src/com/uwb/bt2j/indexer/util/IndexTypes.java
@@ -4,8 +4,8 @@ public class IndexTypes {
   public static String gEbwt_ext;
 
   //64-bit constants
-  public static final long OFF_MASK     = 0xFFFFFFFF;
-  public static final long OFF_LEN_MASK = 0xC0000000;
-  public static final long LS_SIZE      = 0x1000000;
-  public static final byte OFF_SIZE     = 8;
+  public static final int OFF_MASK     = 0xFFFFFFFF;
+  public static final int OFF_LEN_MASK = 0xC0000000;
+  public static final int LS_SIZE      = 0x1000000;
+  public static final int OFF_SIZE     = 8;
 }
